initial-commit
Create the initial version
a. Create the package cse360assign3.
b. Copy the file Calculator.java to this package. Document this file
using the
Javadoc comment style but do not make any code changes to the file at
this time.
c. Generate the Junit tests for this code. There should be seven test
stub methods
generated, one for each method. All but the testCalculator method
should include the default fail assert. The testCalculator method
should test to make sure that the constructor does not return a null.
d. Commit the project to git and github