second-commit
a. Make the following changes to the Calculator class.
i. The add method should add the parameter to the total variable
ii. The subtract method should subtract the parameter from the total
variable iii. The getTotal method should return the correct total
iv. The multiply method should multiply the total by the total by the
parameter
v. The divide method should divide the total by the parameter. Use
integer
division. If the parameter is zero, set the total to zero. Do not print
an
error message.
vi. The getHistory method should NOT be changed at this time and should
continue to return an empty String
b. Update the CalculatorTest to test all methods. Be sure to test for
divide by zero. c. Commit the project to git and github.